+++ lib/geophrase_flutter.dart
@@ -1,4 +1,2 @@
-library geophrase_flutter;
-
 export 'src/geophrase_types.dart';
 export 'src/geophrase_connect.dart';